arranging the test cases into different files Refactoring #215
